clean up and readded td_error + Base Agent for DQN
