--- source/ppo_imitation.py
@@ -1,809 +1,0 @@
-# File for imitation learning on the new basic.wad to compare it to basic reinforcement learning
-
-from __future__ import print_function
-
-from time import sleep
-import vizdoom as vzd
-from argparse import ArgumentParser
-
-DEFAULT_CONFIG = "lib/ViZDoom/scenarios/basic.cfg"
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-import torch.optim as optim
-from torch.distributions import Categorical
-from torch import manual_seed
-from torch import from_numpy, no_grad, save, load, tensor, clamp
-from torch import float as torch_float
-from torch import long as torch_long
-from torch import min as torch_min
-from torch.utils.data.sampler import BatchSampler, SubsetRandomSampler
-
-import numpy as np
-import random
-import itertools as it
-import skimage.transform
-
-from vizdoom import Mode
-from time import sleep, time
-from collections import deque, namedtuple
-from tqdm import trange
-
-from utilities import normalizeToRange, plotData
-
-# Q-learning settings
-learning_rate = 0.00025
-discount_factor = 0.99
-train_epochs = 5
-learning_steps_per_epoch = 2000
-replay_memory_size = 10000
-
-# Tell the agent to get actions in a supervised manner
-supervised = True
-
-# NN learning settings
-batch_size = 64
-
-# Training regime
-test_episodes_per_epoch = 10#0
-
-# Other parameters
-frame_repeat = 12
-resolution = (30, 45)
-episodes_to_watch = 10
-
-model_savefile = "./model-doom.pth"
-save_model = True
-load_model = False
-skip_learning = False
-
-# Configuration file path
-# config_file_path = "../../scenarios/simpler_basic.cfg"
-# config_file_path = "../../scenarios/rocket_basic.cfg"
-config_file_path = DEFAULT_CONFIG
-
-# Uses GPU if available
-if torch.cuda.is_available():
-    DEVICE = torch.device('cuda')
-    torch.backends.cudnn.benchmark = True
-else:
-    DEVICE = torch.device('cpu')
-
-
-def preprocess(img):
-    """Down samples image to resolution"""
-    img = skimage.transform.resize(img, resolution)
-    img = img.astype(np.float32)
-    img = np.expand_dims(img, axis=0)
-    return img
-
-def create_simple_game():
-    print("Initializing doom...")
-    game = vzd.DoomGame()
-    game.load_config(config_file_path)
-
-    # Enables freelook in engine
-    # game.add_game_args("+freelook 1")
-
-    # Enables spectator mode, so you can play. Sounds strange but it is the agent who is supposed to watch not you.
-    game.set_window_visible(True)
-    game.set_mode(vzd.Mode.SPECTATOR)
-    game.set_screen_format(vzd.ScreenFormat.RGB24)
-    game.set_screen_resolution(vzd.ScreenResolution.RES_640X480)
-
-    game.init()
-    print("Doom initialized.")
-
-    return game
-
-def test(game, agent):
-    """Runs a test_episodes_per_epoch episodes and prints the result"""
-    print("\nTesting...")
-    test_scores = []
-    for test_episode in trange(test_episodes_per_epoch, leave=False):
-        game.new_episode()
-        while not game.is_episode_finished():
-            state = preprocess(game.get_state().screen_buffer)
-            best_action_index = agent.get_action(state)
-
-            game.make_action(actions[best_action_index], frame_repeat)
-        r = game.get_total_reward()
-        test_scores.append(r)
-
-    test_scores = np.array(test_scores)
-    print("Results: mean: %.1f +/- %.1f," % (
-        test_scores.mean(), test_scores.std()), "min: %.1f" % test_scores.min(),
-          "max: %.1f" % test_scores.max())
-
-def run(game, agent, actions, num_epochs, frame_repeat, steps_per_epoch=2000):
-    """
-    Run num epochs of training episodes.
-    Skip frame_repeat number of frames after each action.
-    """
-
-    start_time = time()
-
-    for epoch in range(num_epochs):
-        game.new_episode()
-        train_scores = []
-        global_step = 0
-        print("\nEpoch #" + str(epoch + 1))
-
-        for _ in trange(steps_per_epoch, leave=False):
-
-            # One episode is 1 action performed in the current state and the reward given
-            state = preprocess(game.get_state().screen_buffer)
-
-            # Get the supervised action here
-            #TODO: Make sure every state gets a corresponding action somehow as transmitted in actions
-            game.advance_action()
-            last_action = game.get_last_action()
-            reward = game.get_last_reward()
-
-            map_supervised_action(last_action, actions)
-
-            print("Action:", last_action)
-            print("Reward:", reward)
-            print("=====================")
-
-            #reward = game.make_action(actions[action], frame_repeat)
-            done = game.is_episode_finished()
-
-            if not done:
-                next_state = preprocess(game.get_state().screen_buffer)
-            else:
-                next_state = np.zeros((1, 30, 45)).astype(np.float32)
-
-            agent.append_memory(state, last_action, reward, next_state, done)
-
-            # performed actions are given as training examples to the agent after one batch is finished
-            if global_step > agent.batch_size:
-                #TODO: Bugfix
-                agent.train()
-
-            # finished the episode by completing the target
-            if done:
-                train_scores.append(game.get_total_reward())
-                game.new_episode()
-
-            global_step += 1
-        
-        print("Episode finished!")
-        print("Total reward:", game.get_total_reward())
-        print("************************")
-        sleep(2.0)
-
-        agent.update_target_net()
-        train_scores = np.array(train_scores)
-
-        print("Results: mean: %.1f +/- %.1f," % (train_scores.mean(), train_scores.std()),
-              "min: %.1f," % train_scores.min(), "max: %.1f," % train_scores.max())
-
-        test(game, agent)
-        if save_model:
-            print("Saving the network weights to:", model_savefile)
-            torch.save(agent.q_net, model_savefile)
-        print("Total elapsed time: %.2f minutes" % ((time() - start_time) / 60.0))
-
-    game.close()
-    return agent, game
-
-def map_supervised_action(last_action, actions):
-    # This function takes the recorded action and maps it to a actions combination for the DuelQ Agent
-    pass
-
-# old if __name__==__main__ of learning_pytorch.py
-def result():    
-
-    # Reinitialize the game with window visible
-    game.close()
-    # game.set_window_visible(True)
-    game.set_mode(Mode.ASYNC_PLAYER)
-    game.init()
-
-    for _ in range(episodes_to_watch):
-        game.new_episode()
-        while not game.is_episode_finished():
-            state = preprocess(game.get_state().screen_buffer)
-            best_action_index = agent.get_action(state)
-
-            # Instead of make_action(a, frame_repeat) in order to make the animation smooth
-            game.set_action(actions[best_action_index])
-            for _ in range(frame_repeat):
-                game.advance_action()
-
-        # Sleep between episodes
-        sleep(1.0)
-        score = game.get_total_reward()
-        print("Total score: ", score)
-
-class DuelQNet(nn.Module):
-    """
-    This is Duel DQN architecture.
-    see https://arxiv.org/abs/1511.06581 for more information.
-    """
-
-    def __init__(self, available_actions_count):
-        super(DuelQNet, self).__init__()
-        self.conv1 = nn.Sequential(
-            nn.Conv2d(1, 8, kernel_size=3, stride=2, bias=False),
-            nn.BatchNorm2d(8),
-            nn.ReLU()
-        )
-
-        self.conv2 = nn.Sequential(
-            nn.Conv2d(8, 8, kernel_size=3, stride=2, bias=False),
-            nn.BatchNorm2d(8),
-            nn.ReLU()
-        )
-
-        self.conv3 = nn.Sequential(
-            nn.Conv2d(8, 8, kernel_size=3, stride=1, bias=False),
-            nn.BatchNorm2d(8),
-            nn.ReLU()
-        )
-
-        self.conv4 = nn.Sequential(
-            nn.Conv2d(8, 16, kernel_size=3, stride=1, bias=False),
-            nn.BatchNorm2d(16),
-            nn.ReLU()
-        )
-
-        self.state_fc = nn.Sequential(
-            nn.Linear(96, 64),
-            nn.ReLU(),
-            nn.Linear(64, 1)
-        )
-
-        self.advantage_fc = nn.Sequential(
-            nn.Linear(96, 64),
-            nn.ReLU(),
-            nn.Linear(64, available_actions_count)
-        )
-
-    def forward(self, x):
-        x = self.conv1(x)
-        x = self.conv2(x)
-        x = self.conv3(x)
-        x = self.conv4(x)
-        x = x.view(-1, 192)
-        x1 = x[:, :96]  # input for the net to calculate the state value
-        x2 = x[:, 96:]  # relative advantage of actions in the state
-        state_value = self.state_fc(x1).reshape(-1, 1)
-        advantage_values = self.advantage_fc(x2)
-        x = state_value + (advantage_values - advantage_values.mean(dim=1).reshape(-1, 1))
-
-        return x
-
-class DQNAgent:
-    def __init__(self, action_size, memory_size, batch_size, discount_factor, 
-                 lr, load_model, epsilon=1, epsilon_decay=0.9996, epsilon_min=0.1):
-        self.action_size = action_size
-        self.epsilon = epsilon
-        self.epsilon_decay = epsilon_decay
-        self.epsilon_min = epsilon_min
-        self.batch_size = batch_size
-        self.discount = discount_factor
-        self.lr = lr
-        self.memory = deque(maxlen=memory_size)
-        self.criterion = nn.MSELoss()
-
-        if load_model:
-            print("Loading model from: ", model_savefile)
-            self.q_net = torch.load(model_savefile)
-            self.target_net = torch.load(model_savefile)
-            self.epsilon = self.epsilon_min
-
-        else:
-            print("Initializing new model")
-            self.q_net = DuelQNet(action_size).to(DEVICE)
-            self.target_net = DuelQNet(action_size).to(DEVICE)
-
-        self.opt = optim.SGD(self.q_net.parameters(), lr=self.lr)
-
-    def get_action(self, state):
-
-        if np.random.uniform() < self.epsilon:
-            return random.choice(range(self.action_size))
-        else:
-            state = np.expand_dims(state, axis=0)
-            state = torch.from_numpy(state).float().to(DEVICE)
-            action = torch.argmax(self.q_net(state)).item()
-            return action
-
-    def update_target_net(self):
-        self.target_net.load_state_dict(self.q_net.state_dict())
-
-    def append_memory(self, state, action, reward, next_state, done):
-        self.memory.append((state, action, reward, next_state, done))
-
-    def train(self):
-        batch = random.sample(self.memory, self.batch_size)
-        batch = np.array(batch, dtype=object)
-
-        states = np.stack(batch[:, 0]).astype(float)
-
-        actions = self.remap_actions(batch[:, 1])
-        # actions = batch[:, 1].astype(int)
-
-        rewards = batch[:, 2].astype(float)
-        with open('batch_stack.txt', 'w') as f:
-            f.write(str(batch[:, 3]))
-        print(batch[:, 3].shape)
-        next_states = np.stack(batch[:, 3]).astype(float)
-        dones = batch[:, 4].astype(bool)
-        not_dones = ~dones
-
-        row_idx = np.arange(self.batch_size)  # used for indexing the batch
-
-        # value of the next states with double q learning
-        # see https://arxiv.org/abs/1509.06461 for more information on double q learning
-        with torch.no_grad():
-            next_states = torch.from_numpy(next_states).float().to(DEVICE)
-            idx = row_idx, np.argmax(self.q_net(next_states).cpu().data.numpy(), 1)
-            next_state_values = self.target_net(next_states).cpu().data.numpy()[idx]
-            next_state_values = next_state_values[not_dones]
-
-        # this defines y = r + discount * max_a q(s', a)
-        q_targets = rewards.copy()
-        q_targets[not_dones] += self.discount * next_state_values
-        q_targets = torch.from_numpy(q_targets).float().to(DEVICE)
-
-        # this selects only the q values of the actions taken
-        idx = row_idx, actions
-        states = torch.from_numpy(states).float().to(DEVICE)
-        action_values = self.q_net(states)[idx].float().to(DEVICE)
-
-        self.opt.zero_grad()
-        td_error = self.criterion(q_targets, action_values)
-        td_error.backward()
-        self.opt.step()
-
-        if self.epsilon > self.epsilon_min:
-            self.epsilon *= self.epsilon_decay
-        else:
-            self.epsilon = self.epsilon_min
-    
-    def remap_actions(self, actions):
-
-        dic = {
-            tuple(list([0, 0, 0])) : 0,
-            tuple(list([0, 0, 1])) : 1,
-            tuple(list([0, 1, 0])) : 2,
-            tuple(list([0, 1, 1])) : 3,
-            tuple(list([1, 0, 0])) : 4,
-            tuple(list([1, 0, 1])) : 5,
-            tuple(list([1, 1, 0])) : 6,
-            tuple(list([1, 1, 1])) : 7
-        }
-
-        actions = [dic[tuple(action)] for action in actions]
-
-        return np.array(actions).astype(int)
-
-
-if __name__ == "__main__":
-    parser = ArgumentParser("PPO with Imitation Learning")
-    parser.add_argument(dest="config",
-                        default=DEFAULT_CONFIG,
-                        nargs="?",
-                        help="Path to the configuration file of the scenario."
-                             " Please see "
-                             "../../scenarios/*cfg for more scenarios.")
-    args = parser.parse_args()
-    #game = vzd.DoomGame()
-
-    # Initialize game and actions
-    game = create_simple_game()
-    n = game.get_available_buttons_size()
-    actions = [list(a) for a in it.product([0, 1], repeat=n)]
-
-    print(actions)
-
-    # Initialize our agent with the set parameters
-    """
-    agent = DQNAgent(len(actions), lr=learning_rate, batch_size=batch_size,
-                     memory_size=replay_memory_size, discount_factor=discount_factor,
-                     load_model=load_model)
-    """
-    agent = PPOAgent()
-
-    episodes = 10
-
-    # Run the training for the set number of epochs
-    if not skip_learning:
-        agent, game = run(game, agent, actions, num_epochs=train_epochs, frame_repeat=frame_repeat,
-                          steps_per_epoch=learning_steps_per_epoch)
-
-        print("======================================")
-        print("Training finished. It's time to watch!")
-
-    game.close()
-
-
-
-
-"""
----------------------------------------------------------------------------------------------------
-
----------------------------------------------------------------------------------------------------
-
----------------------------------------------------------------------------------------------------
-"""
-
-Transition = namedtuple('Transition', ['state', 'action', 'a_log_prob', 'reward', 'next_state'])
-
-
-class PPOAgent:
-    """
-    PPOAgent implements the PPO RL algorithm (https://arxiv.org/abs/1707.06347).
-    It works with a set of discrete actions.
-    It uses the Actor and Critic neural network classes defined below.
-    """
-
-    def __init__(self, numberOfInputs, numberOfActorOutputs, clip_param=0.2, max_grad_norm=0.5, ppo_update_iters=5,
-                 batch_size=8, gamma=0.99, use_cuda=False, actor_lr=0.001, critic_lr=0.003, seed=None):
-        super().__init__()
-        if seed is not None:
-            manual_seed(seed)
-
-        # Hyper-parameters
-        self.clip_param = clip_param
-        self.max_grad_norm = max_grad_norm
-        self.ppo_update_iters = ppo_update_iters
-        self.batch_size = batch_size
-        self.gamma = gamma
-        self.use_cuda = use_cuda
-
-        # models
-        self.actor_net = Actor(numberOfInputs, numberOfActorOutputs)
-        self.critic_net = Critic(numberOfInputs)
-
-        if self.use_cuda:
-            self.actor_net.cuda()
-            self.critic_net.cuda()
-
-        # Create the optimizers
-        self.actor_optimizer = optim.Adam(self.actor_net.parameters(), actor_lr)
-        self.critic_net_optimizer = optim.Adam(self.critic_net.parameters(), critic_lr)
-
-        # Training stats
-        self.buffer = []
-
-    def work(self, agentInput, type_="selectAction"):
-        """
-        Forward pass of the PPO agent. Depending on the type_ argument, it either explores by sampling its actor's
-        softmax output, or eliminates exploring by selecting the action with the maximum probability (argmax).
-
-        :param agentInput: The actor neural network input vector
-        :type agentInput: vector
-        :param type_: "selectAction" or "selectActionMax", defaults to "selectAction"
-        :type type_: str, optional
-        """
-        agentInput = from_numpy(np.array(agentInput)).float().unsqueeze(0)  # Add batch dimension with unsqueeze
-
-        if self.use_cuda:
-            agentInput = agentInput.cuda()
-
-        with no_grad():
-            action_prob = self.actor_net(agentInput)
-
-        if type_ == "selectAction":
-            c = Categorical(action_prob)
-            action = c.sample()
-            return action.item(), action_prob[:, action.item()].item()
-        elif type_ == "selectActionMax":
-            return np.argmax(action_prob).item(), 1.0
-
-    def save(self, path):
-        """
-        Save actor and critic models in the path provided.
-
-        :param path: path to save the models
-        :type path: str
-        """
-        save(self.actor_net.state_dict(), path + '_actor.pkl')
-        save(self.critic_net.state_dict(), path + '_critic.pkl')
-
-    def load(self, path):
-        """
-        Load actor and critic models from the path provided.
-
-        :param path: path where the models are saved
-        :type path: str
-        """
-        actor_state_dict = load(path + '_actor.pkl')
-        critic_state_dict = load(path + '_critic.pkl')
-        self.actor_net.load_state_dict(actor_state_dict)
-        self.critic_net.load_state_dict(critic_state_dict)
-
-    def storeTransition(self, transition):
-        """
-        Stores a transition in the buffer to be used later.
-
-        :param transition: contains state, action, action_prob, reward, next_state
-        :type transition: namedtuple('Transition', ['state', 'action', 'a_log_prob', 'reward', 'next_state'])
-        """
-        self.buffer.append(transition)
-
-    def trainStep(self, batchSize=None):
-        """
-        Performs a training step for the actor and critic models, based on transitions gathered in the
-        buffer. It then resets the buffer.
-
-        :param batchSize: Overrides agent set batch size, defaults to None
-        :type batchSize: int, optional
-        """
-        # Default behaviour waits for buffer to collect at least one batch_size of transitions
-        if batchSize is None:
-            if len(self.buffer) < self.batch_size:
-                return
-            batchSize = self.batch_size
-
-        # Extract states, actions, rewards and action probabilities from transitions in buffer
-        state = tensor([t.state for t in self.buffer], dtype=torch_float)
-        action = tensor([t.action for t in self.buffer], dtype=torch_long).view(-1, 1)
-        reward = [t.reward for t in self.buffer]
-        old_action_log_prob = tensor([t.a_log_prob for t in self.buffer], dtype=torch_float).view(-1, 1)
-
-        # Unroll rewards
-        R = 0
-        Gt = []
-        for r in reward[::-1]:
-            R = r + self.gamma * R
-            Gt.insert(0, R)
-        Gt = tensor(Gt, dtype=torch_float)
-
-        # Send everything to cuda if used
-        if self.use_cuda:
-            state, action, old_action_log_prob = state.cuda(), action.cuda(), old_action_log_prob.cuda()
-            Gt = Gt.cuda()
-
-        # Repeat the update procedure for ppo_update_iters
-        for i in range(self.ppo_update_iters):
-            # Create randomly ordered batches of size batchSize from buffer
-            for index in BatchSampler(SubsetRandomSampler(range(len(self.buffer))), batchSize, False):
-                # Calculate the advantage at each step
-                Gt_index = Gt[index].view(-1, 1)
-                V = self.critic_net(state[index])
-                delta = Gt_index - V
-                advantage = delta.detach()
-
-                # Get the current probabilities
-                # Apply past actions with .gather()
-                action_prob = self.actor_net(state[index]).gather(1, action[index])  # new policy
-
-                # PPO
-                ratio = (action_prob / old_action_log_prob[index])  # Ratio between current and old policy probabilities
-                surr1 = ratio * advantage
-                surr2 = clamp(ratio, 1 - self.clip_param, 1 + self.clip_param) * advantage
-
-                # update actor network
-                action_loss = -torch_min(surr1, surr2).mean()  # MAX->MIN descent
-                self.actor_optimizer.zero_grad()  # Delete old gradients
-                action_loss.backward()  # Perform backward step to compute new gradients
-                nn.utils.clip_grad_norm_(self.actor_net.parameters(), self.max_grad_norm)  # Clip gradients
-                self.actor_optimizer.step()  # Perform training step based on gradients
-
-                # update critic network
-                value_loss = F.mse_loss(Gt_index, V)
-                self.critic_net_optimizer.zero_grad()
-                value_loss.backward()
-                nn.utils.clip_grad_norm_(self.critic_net.parameters(), self.max_grad_norm)
-                self.critic_net_optimizer.step()
-
-        # After each training step, the buffer is cleared
-        del self.buffer[:]
-
-
-class Actor(nn.Module):
-    def __init__(self, numberOfInputs, numberOfOutputs):
-        super(Actor, self).__init__()
-        self.fc1 = nn.Linear(numberOfInputs, 10)
-        self.fc2 = nn.Linear(10, 10)
-        self.action_head = nn.Linear(10, numberOfOutputs)
-
-    def forward(self, x):
-        x = F.relu(self.fc1(x))
-        x = F.relu(self.fc2(x))
-        action_prob = F.softmax(self.action_head(x), dim=1)
-        return action_prob
-
-
-class Critic(nn.Module):
-    def __init__(self, numberOfInputs):
-        super(Critic, self).__init__()
-        self.fc1 = nn.Linear(numberOfInputs, 10)
-        self.fc2 = nn.Linear(10, 10)
-        self.state_value = nn.Linear(10, 1)
-
-    def forward(self, x):
-        x = F.relu(self.fc1(x))
-        x = F.relu(self.fc2(x))
-        value = self.state_value(x)
-        return value
-
-
-
-
-"""
----------------------------------------------------------------------------------------------------
-
----------------------------------------------------------------------------------------------------
-
----------------------------------------------------------------------------------------------------
-"""
-
-class CartpoleRobot(RobotSupervisor):
-    def __init__(self):
-        super().__init__()
-        
-        self.observation_space = Box(low=np.array([-0.4, -np.inf, -1.3, -np.inf]),
-                                     high=np.array([0.4, np.inf, 1.3, np.inf]),
-                                     dtype=np.float64)
-        self.action_space = Discrete(2)
-
-        self.robot = self.getSelf()  # Grab the robot reference from the supervisor to access various robot methods
-        self.positionSensor = self.getDevice("polePosSensor")
-        self.positionSensor.enable(self.timestep)
-
-        self.poleEndpoint = self.getFromDef("POLE_ENDPOINT")
-        self.wheels = []
-        for wheelName in ['wheel1', 'wheel2', 'wheel3', 'wheel4']:
-            wheel = self.getDevice(wheelName)  # Get the wheel handle
-            wheel.setPosition(float('inf'))  # Set starting position
-            wheel.setVelocity(0.0)  # Zero out starting velocity
-            self.wheels.append(wheel)
-
-        self.stepsPerEpisode = 200  # Max number of steps per episode
-        self.episodeScore = 0  # Score accumulated during an episode
-        self.episodeScoreList = []  # A list to save all the episode scores, used to check if task is solved
-
-    def get_observations(self):
-        # Position on z axis
-        cartPosition = normalizeToRange(self.robot.getPosition()[2], -0.4, 0.4, -1.0, 1.0)
-        # Linear velocity on z axis
-        cartVelocity = normalizeToRange(self.robot.getVelocity()[2], -0.2, 0.2, -1.0, 1.0, clip=True)
-        # Pole angle off vertical
-        poleAngle = normalizeToRange(self.positionSensor.getValue(), -0.23, 0.23, -1.0, 1.0, clip=True)
-        # Angular velocity x of endpoint
-        endpointVelocity = normalizeToRange(self.poleEndpoint.getVelocity()[3], -1.5, 1.5, -1.0, 1.0, clip=True)
-
-        return [cartPosition, cartVelocity, poleAngle, endpointVelocity]
-
-    def get_reward(self, action=None):
-        return 1
-
-    def is_done(self):
-        if self.episodeScore > 195.0:
-            return True
-
-        poleAngle = round(self.positionSensor.getValue(), 2)
-        if abs(poleAngle) > 0.261799388:  # 15 degrees off vertical
-            return True
-
-        cartPosition = round(self.robot.getPosition()[2], 2)  # Position on z axis
-        if abs(cartPosition) > 0.39:
-            return True
-        
-        return False
-        
-    def solved(self):
-        if len(self.episodeScoreList) > 100:  # Over 100 trials thus far
-            if np.mean(self.episodeScoreList[-100:]) > 195.0:  # Last 100 episodes' scores average value
-                return True
-        return False
-
-    def get_default_observation(self):
-        return [0.0 for _ in range(self.observation_space.shape[0])]
-
-    def apply_action(self, action):
-        action = int(action[0])
-
-        if action == 0:
-            motorSpeed = 5.0
-        else:
-            motorSpeed = -5.0
-
-        for i in range(len(self.wheels)):
-            self.wheels[i].setPosition(float('inf'))
-            self.wheels[i].setVelocity(motorSpeed)
-
-    def render(self, mode='human'):
-        print("render() is not used")
-
-    def get_info(self):
-        return None
-
-
-env = CartpoleRobot()
-
-if True:
-    agent = PPOAgent(numberOfInputs=env.observation_space.shape[0], numberOfActorOutputs=env.action_space.n)
-
-    solved = False
-    episodeCount = 0
-    episodeLimit = 2000
-
-    # Run outer loop until the episodes limit is reached or the task is solved
-    while not solved and episodeCount < episodeLimit:
-        observation = env.reset()  # Reset robot and get starting observation
-        env.episodeScore = 0
-
-        for step in range(env.stepsPerEpisode):
-            # In training mode the agent samples from the probability distribution, naturally implementing exploration
-            selectedAction, actionProb = agent.work(observation, type_="selectAction")
-            # Step the supervisor to get the current selectedAction's reward, the new observation and whether we reached 
-            # the done condition
-            newObservation, reward, done, info = env.step([selectedAction])
-
-            # Save the current state transition in agent's memory
-            trans = Transition(observation, selectedAction, actionProb, reward, newObservation)
-            agent.storeTransition(trans)
-            if done:
-                # Save the episode's score
-                env.episodeScoreList.append(env.episodeScore)
-                agent.trainStep(batchSize=step)
-                solved = env.solved()  # Check whether the task is solved
-                break
-
-            env.episodeScore += reward  # Accumulate episode reward
-            observation = newObservation  # observation for next step is current step's newObservation
-        
-        print("Episode #", episodeCount, "score:", env.episodeScore)
-        episodeCount += 1  # Increment episode counter
-
-    if not solved:
-        print("Task is not solved, deploying agent for testing...")
-    elif solved:
-        print("Task is solved, deploying agent for testing...")
-    observation = env.reset()
-    while True:
-        selectedAction, actionProb = agent.work(observation, type_="selectActionMax")
-        observation, _, _, _ = env.step([selectedAction])
-        
-else:
-    agent = DQNAgent(DEVICE, numberOfInputs=env.observation_space.shape[0], numberOfActorOutputs=env.action_space.n)
-    
-    solved = False
-    episodeCount = 0
-    episodeLimit = 2000
-
-    # Run outer loop until the episodes limit is reached or the task is solved
-    while not solved and episodeCount < episodeLimit:
-        observation = env.reset()  # Reset robot and get starting observation
-        env.episodeScore = 0
-
-        for step in range(env.stepsPerEpisode):
-            # In training mode the agent samples from the probability distribution, naturally implementing exploration
-            selectedAction = agent.work(observation)
-            # Step the supervisor to get the current selectedAction's reward, the new observation and whether we reached 
-            # the done condition
-            newObservation, reward, done, info = env.step([selectedAction])
-
-            # Save the current state transition in agent's memory
-            trans = Transition(observation, selectedAction, reward, newObservation)
-            agent.storeTransition(trans)
-            if done:
-                # Save the episode's score
-                env.episodeScoreList.append(env.episodeScore)
-                agent.trainStep(batchSize=step)
-                solved = env.solved()  # Check whether the task is solved
-                break
-
-            env.episodeScore += reward  # Accumulate episode reward
-            observation = newObservation  # observation for next step is current step's newObservation
-        
-        print("Episode #", episodeCount, "score:", env.episodeScore)
-        episodeCount += 1  # Increment episode counter
-
-    if not solved:
-        print("Task is not solved, deploying agent for testing...")
-    elif solved:
-        print("Task is solved, deploying agent for testing...")
-    observation = env.reset()
-    while True:
-        selectedAction, actionProb = agent.work(observation, type_="selectActionMax")
-        observation, _, _, _ = env.step([selectedAction])
-
